Add Export button to content edit pages
Add an export button alongside save and publish on content edit pages.

Add a new export single route that will export a single item by ID and
give the export download file as a response.

Add styles for styling the export button the same way as the current
content buttons.